fix(bl31): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall be assigned with an essential type
category.

State: waiting
Link: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/40540

+++ include/bl31/interrupt_mgmt.h
@@ -149,8 +149,8 @@ int32_t register_interrupt_type_handler(uint32_t type,
 					interrupt_type_handler_t handler,
 					uint32_t flags);
 interrupt_type_handler_t get_interrupt_type_handler(uint32_t type);
-int disable_intr_rm_local(uint32_t type, uint32_t security_state);
-int enable_intr_rm_local(uint32_t type, uint32_t security_state);
+int disable_intr_rm_local(uint32_t type, size_t security_state);
+int enable_intr_rm_local(uint32_t type, size_t security_state);
 
 #endif /*__ASSEMBLER__*/
 #endif /* INTERRUPT_MGMT_H */
